+++ corrupted_clock/src/cli_args/column_show_arg.rs
@@ -1,0 +1,18 @@
+use clap::Args;
+
+use crate::AtLeastOne;
+
+#[derive(Debug, Args)]
+pub struct ColumnShowArg {
+    #[arg(short, long)]
+    /// How many colums are printed within a row at max
+    ///
+    /// Note: Must be greater than zero
+    colums_num: Option<AtLeastOne>,
+}
+
+impl ColumnShowArg {
+    pub fn colums_num(&self) -> Option<AtLeastOne> {
+        self.colums_num
+    }
+}